backend/requirements.txt
# Python Dependencies for Sensor Reading Service
pymodbus==3.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0

